# openapi.yaml
...
openapi: 3.0.3
info:
  title: Medium Public Unofficial API
  description: >-
    
    This is an unofficial Medium API used to fetch public data such as

    users, articles, lists, etc...





    This API is heavily inspired by [Nishu Jain's Medium API](https://mediumapi.com/), in fact it's nearly the same, I just wanted to make it open source and free.





    It is done using Medium's "official" API - https://medium.com/_/api (I quote official because there is no documentation for this API and no way of knowing it exists without using Developer Tools in the web browser) - and some web scraping.





    Useful links:



    - [API's repository](https://github.com/estebanthi/medium-api)



    - [OpenAPI specification](https://github.com/estebanthi/medium-api/blob/master/openapi.yaml)
  contact:
    email: contact@esteban-thilliez.com
  version: "1.0"
servers:
  - url: https://mediumapi.esteban-thilliez.com/
...
paths:
  "/users/id-for/{username}":
    get:
      summary: Get User ID
      tags:
        - Users
      description: >
        
        Returns the unique `userId` for the given `username`.





        **Note:** You can find the `username` from the user's/author's profile page URL.



        - **username**.medium.com



        - medium.com/@**username**
      parameters:
        - name: username
          in: path
          required: true
          description: It's a unique string chosen by every single Medium user.
          schema:
            type: string
            example: estebanthi
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: string
                    description: Unique hash id of the user.
                    example: 24babe9bccb4
  "/users/{userId}":
    get:
      summary: Get User Info
      tags:
        - Users
      description: >-
        Returns user related information like *Full name, Bio, Followers

        count, Following count, Twitter username, Profile-image URL, User ID, etc...*
      parameters:
        - name: userId
          in: path
          required: true
          description: It's a unique hash id assigned to every single Medium user.
          schema:
            type: string
            example: 24babe9bccb4
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      userId:
                        type: string
                        example: 24babe9bccb4
                      username:
                        type: string
                        example: estebanthi
                      fullName:
                        type: string
                        example: Esteban Thilliez
                      bio:
                        type: string
                        example: "I’m Esteban, and I enjoy writing about programming, trading,
                          productivity, knowledge management, books and movies!
                          Contact me here: contact@esteban-thilliez.com"
                      avatar:
                        type: string
                        example: https://miro.medium.com/1*sPkU-dWfO_U6_Om7knbXtQ.jpeg
                      banner:
                        type: string
                        example:
                      followersCount:
                        type: integer
                        format: int32
                        example: 1001
                      followingCount:
                        type: integer
                        format: int32
                        example: 6
                      twitter:
                        type: string
                        example: null
                      isCreatorPartnerProgramEnrolled:
                        type: boolean
                        example: true
                      topWriterIn:
                        type: array
                        items:
                          type: string
                          example: productivity
                      joinedDate:
                        type: string
                        example: 2021-06-17T07:11:06.041Z
  "/users/{userId}/following":
    get:
      summary: Get User Following
      tags:
        - Users
      description: Returns the list of `userIds` of the user's followings.
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique hash id assigned to every single Medium user.
          schema:
            type: string
            example: 24babe9bccb4
        - name: count
          in: query
          required: false
          description: To limit the number of results.
          schema:
            type: integer
            example: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: string
                      example: 27f2a48b3308
  "/users/{userId}/followers":
    get:
      summary: Get User Followers
      tags:
        - Users
      description: |
        
        Returns the list of `userIds` of the user's followers.
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique hash id assigned to every single Medium user.
          schema:
            type: string
            example: 24babe9bccb4
        - name: count
          in: query
          required: false
          description: To limit the number of results.
          schema:
            type: integer
            example: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: string
                      example: 1044fc419f11
  "/users/{userId}/articles":
    get:
      summary: Get User's Articles
      tags:
        - Users
      description: Returns the list of articles (`articleIds`) written by the user.
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique hash id assigned to every single Medium user.
          schema:
            type: string
            example: 24babe9bccb4
        - name: count
          in: query
          required: false
          description: To limit the number of results.
          schema:
            type: integer
            example: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: string
                      example: 2cadf5e2561e
  "/users/{userId}/publications":
    get:
      summary: Get User's Publications
      tags:
        - Users
      description: |-
        Returns the list of publications (`publicationIds`) where the user
        is the editor.
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique hash id assigned to every single Medium user.
          schema:
            type: string
            example: 24babe9bccb4
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: string
                      example: a648dc4ecb66
  "/users/{userId}/top-articles":
    get:
      summary: Get User's Top Articles
      tags:
        - Users
      description: |-
        Returns a list of `articleIds` of user's articles, sorted by number
        of claps, for a given `userId`.
        
        
        **Note:** If `count` is too high, it may take a long time/crash depending on the server.
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique hash id assigned to every single Medium user.
          schema:
            type: string
            example: 24babe9bccb4
        - name: count
          in: query
          required: false
          description: To limit the number of results.
          schema:
            type: integer
            example: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: string
                      example: 03hO6bnd6jap
  "/users/{userId}/lists":
    get:
      summary: Get User's Lists
      tags:
        - Users
      description: Returns the list of `listIds` of the user's lists.
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique hash id assigned to every single Medium user.
          schema:
            type: string
            example: 24babe9bccb4
        - name: count
          in: query
          required: false
          description: To limit the number of results.
          schema:
            type: integer
            example: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: string
                      example: 1f9f2037ef8e
...